requirements.txt
fastapi==0.115.0
pydantic==2.9.2
openai==1.51.0
python-dotenv==1.0.1
bleach==6.1.0
uvicorn==0.32.0